Refactor reader layout
